Working on the budget actual

+++ Financial-Tracking-API/src/main/java/com/Financial_Tracking_API/Financial_Tracking_API/Expense/ExpenseRepository.java
@@ -2,6 +2,6 @@ package com.Financial_Tracking_API.Financial_Tracking_API.Expense;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ExpenseRepository extends JpaRepository<Expense, Long> {
+public interface ExpenseRepository extends JpaRepository<Expense, Integer> {
 
 }
